+++ cmd/prepare.go
@@ -1,18 +1,18 @@
 package cmd
 
 import (
-	"github.com/broadinstitute/revere/internal"
+	"github.com/broadinstitute/revere/internal/actions"
+	config2 "github.com/broadinstitute/revere/internal/configuration"
 	"github.com/broadinstitute/revere/internal/statuspage"
-	"github.com/broadinstitute/revere/pkg"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
 var prepareCmd = &cobra.Command{
 	Use:   "prepare",
-	Short: "Configure Statuspage.io based on the config",
+	Short: "Configure Statuspage.io based on the configuration",
 	Long: `This command diffs the Statuspage components and groups from
-the config with what's present on the remote, matching
+the configuration with what's present on the remote, matching
 based on name alone. It then sequentially deletes,
 creates, and patches resources such that a subsequent
 diff would identify no changes.`,
@@ -20,10 +20,10 @@ diff would identify no changes.`,
 }
 
 func Prepare(*cobra.Command, []string) {
-	config, err := pkg.AssembleConfig(viper.GetViper())
+	config, err := config2.AssembleConfig(viper.GetViper())
 	cobra.CheckErr(err)
 	client := statuspage.Client(config)
-	err = internal.ReconcileComponents(config, client)
+	err = actions.ReconcileComponents(config, client)
 	cobra.CheckErr(err)
 }
