+++ internal/shared/util.go
@@ -1,9 +1,11 @@
 package shared
 
-import "github.com/broadinstitute/revere/pkg"
+import (
+	"github.com/broadinstitute/revere/internal/configuration"
+)
 
-// LogLn prints each non-empty string if config.Verbose, only the first string otherwise.
-func LogLn(config *pkg.Config, always string, verbose ...string) {
+// LogLn prints each non-empty string if configuration.Verbose, only the first string otherwise.
+func LogLn(config *configuration.Config, always string, verbose ...string) {
 	if always != "" {
 		println(always)
 	}
