+++ internal/statuspage/client.go
@@ -1,12 +1,12 @@
 package statuspage
 
 import (
+	"github.com/broadinstitute/revere/internal/configuration"
 	"github.com/broadinstitute/revere/internal/shared"
-	"github.com/broadinstitute/revere/pkg"
 	"github.com/go-resty/resty/v2"
 )
 
-func Client(config *pkg.Config) *resty.Client {
+func Client(config *configuration.Config) *resty.Client {
 	return shared.BaseClient(config).
 		SetHostURL(config.Statuspage.ApiRoot).
 		SetAuthScheme("OAuth").
